pyproject.toml
[project]
name = "mfdp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.13.1",
    "alembic>=1.13.3",
    "asyncpg>=0.29.0",
    "bcrypt>=4.2.0",
    "catboost>=1.2.7",
    "category-encoders>=2.6.4",
    "click>=8.1.7",
    "fastapi>=0.115.0",
    "google-generativeai>=0.8.3",
    "greenlet>=3.1.1",
    "httpx>=0.27.2",
    "numpy==1.26.4",
    "openai>=1.51.2",
    "pandas>=2.2.3",
    "pgqueuer>=0.9.2",
    "pgvector>=0.3.5",
    "pip>=24.2",
    "pydantic-settings>=2.5.2",
    "pydantic>=2.9.2",
    "python-multipart>=0.0.12",
    "ruff>=0.6.9",
    "sentence-transformers>=3.1.1",
    "uvicorn>=0.32.0",
]

[tool.uv]
dev-dependencies = [
    "bs4>=0.0.2",
    "matplotlib>=3.9.2",
    "pip>=24.2",
    "pyarrow>=17.0.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "seaborn>=0.13.2",
]
